requirements.txt
pip>=24.2
setuptools>=75.2.0
wheel>=0.44.0
aiogram>=3.0.0
loguru>=0.7.3
requests>=2.32.3
aiogram>=3.16.0
python-dotenv>=1.0.1